# src/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Express API Documentation",
    "version": "1.0.0",
    "description": "API documentation for Express TypeScript app"
  },
  "servers": [
    {
      "url": "/api",
      "description": "Main API"
    },
    {
      "url": "/api/v2",
      "description": "API Version 2"
    }
  ],
  "paths": {
    "/": {
      "get": {
        "tags": ["Welcome"],
        "summary": "Welcome Message",
        "description": "Return Welcome Message.",
        "responses": {
          "200": {
            "description": "The requested todo item.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Welcome to ExpressJs Api!"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/todos": {
      "get": {
        "tags": ["Todos"],
        "summary": "Get all todos",
        "parameters": [
          {
            "in": "query",
            "name": "page",
            "schema": {
              "type": "integer"
            },
            "description": "The page number (default is 1)"
          },
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "type": "integer"
            },
            "description": "Number of items per page (default is 10)"
          }
        ],
        "responses": {
          "200": {
            "description": "A list of todos."
          }
        }
      },
      "post": {
        "tags": ["Todos"],
        "summary": "Create a new todo",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "title": {
                    "type": "string",
                    "example": "New task"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Todo created successfully."
          }
        }
      }
    },
    "/todos/{id}": {
      "get": {
        "tags": ["Todos"],
        "summary": "Get a specific todo",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            },
            "description": "ID of the todo to retrieve"
          }
        ],
        "responses": {
          "200": {
            "description": "The requested todo item."
          }
        }
      },
      "patch": {
        "tags": ["Todos"],
        "summary": "Update a todo",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            },
            "description": "ID of the todo to update"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "title": {
                    "type": "string",
                    "example": "Updated task"
                  },
                  "completed": {
                    "type": "boolean",
                    "example": true
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Todo updated successfully."
          }
        }
      },
      "delete": {
        "tags": ["Todos"],
        "summary": "Delete a todo",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            },
            "description": "ID of the todo to delete"
          }
        ],
        "responses": {
          "204": {
            "description": "Todo deleted successfully."
          }
        }
      }
    }
  },
  "components": {},
  "tags": [
    {
      "name": "Welcome",
      "description": "Welcome To Expressjs App"
    },
    {
      "name": "Todos",
      "description": "Todo management"
    }
  ]
}